+++ src/mail-app/settings/SettingsView.ts
@@ -154,7 +154,7 @@ export class SettingsView extends BaseTopLevelView implements TopLevelView<Setti
 							locator.systemFacade,
 							locator.logins.getUserController(),
 							locator.usageTestController,
-							locator.publicKeyProvider,
+							locator.publicIdentityKeyProvider,
 							locator.themeController,
 						)
 						settingsViewer.init()
