+++ test/tests/Suite.ts
@@ -47,7 +47,9 @@ import "./api/worker/facades/MailExportTokenFacadeTest.js"
 import "./api/worker/facades/MailFacadeTest.js"
 import "./api/worker/facades/PQFacadeTest.js"
 import "./api/worker/facades/PQMessageTest.js"
-import "./api/worker/facades/PublicKeyProviderTest.js"
+import "./api/worker/facades/PublicEncryptionKeyProviderTest.js"
+import "./api/worker/facades/PublicIdentityKeyProviderTest.js"
+import "./api/worker/facades/IdentityKeyTrustDatabaseTest.js"
 import "./api/worker/facades/RolloutFacadeTest.js"
 import "./api/worker/facades/RsaPqPerformanceTest.js"
 import "./api/worker/facades/UserFacadeTest.js"
