+++ test/tests/api/worker/facades/KeyVerificationFacadeTest.ts
@@ -1,11 +1,9 @@
 import o from "@tutao/otest"
-import { KeyVerificationFacade, TrustedIdentity } from "../../../../../src/common/api/worker/facades/lazy/KeyVerificationFacade"
-import { SqlCipherFacade } from "../../../../../src/common/native/common/generatedipc/SqlCipherFacade"
+import { KeyVerificationFacade } from "../../../../../src/common/api/worker/facades/lazy/KeyVerificationFacade"
 import { matchers, object, verify, when } from "testdouble"
-import { SqlType, TaggedSqlValue } from "../../../../../src/common/api/worker/offline/SqlValue"
 import { concat, hexToUint8Array, uint8ArrayToHex, Versioned } from "@tutao/tutanota-utils"
-import { IdentityKeySourceOfTrust, EncryptionKeyVerificationState } from "../../../../../src/common/api/common/TutanotaConstants"
-import { bytesToEd25519PublicKey, PublicKey, sha256Hash } from "@tutao/tutanota-crypto"
+import { EncryptionKeyVerificationState, IdentityKeySourceOfTrust, PublicKeyIdentifierType } from "../../../../../src/common/api/common/TutanotaConstants"
+import { bytesToEd25519PublicKey, sha256Hash } from "@tutao/tutanota-crypto"
 import testData from "../crypto/CompatibilityTestData.json"
 import { EncodedEd25519Signature, SigningKeyPairType, SigningPublicKey } from "../../../../../src/common/api/worker/facades/Ed25519Facade"
 import { Mode } from "../../../../../src/common/api/common/Env"
@@ -14,242 +12,166 @@ import { assertThrows } from "@tutao/tutanota-test-utils"
 import { KeyVerificationMismatchError } from "../../../../../src/common/api/common/error/KeyVerificationMismatchError"
 import { PublicKeySignatureFacade } from "../../../../../src/common/api/worker/facades/PublicKeySignatureFacade"
 import { ProgrammingError } from "../../../../../src/common/api/common/error/ProgrammingError"
+import { PublicIdentityKeyProvider } from "../../../../../src/common/api/worker/facades/PublicIdentityKeyProvider"
+import { IdentityKeyTrustDatabase, TrustDBEntry } from "../../../../../src/common/api/worker/facades/IdentityKeyTrustDatabase"
+import { MaybeSignedPublicKey, PublicKeyIdentifier } from "../../../../../src/common/api/worker/facades/PublicEncryptionKeyProvider"
+import { createTestEntity } from "../../../TestUtils"
+import { PublicKeySignatureTypeRef } from "../../../../../src/common/api/entities/sys/TypeRefs"
+import { TaggedSqlValue } from "../../../../../src/common/api/worker/offline/SqlValue"
 
 const { anything } = matchers
 
 const PUBLIC_KEY_BYTES = hexToUint8Array(testData.ed25519Tests[0].alicePublicKeyHex)
 const PUBLIC_KEY = bytesToEd25519PublicKey(PUBLIC_KEY_BYTES)
 const PUBLIC_KEY_FINGERPRINT = uint8ArrayToHex(sha256Hash(concat(new Uint8Array([0]), new Uint8Array([SigningKeyPairType.Ed25519]), PUBLIC_KEY_BYTES)))
-const PUBLIC_KEY_TRUSTED_IDENTITY: TrustedIdentity = {
-	fingerprint: PUBLIC_KEY_FINGERPRINT,
-	publicIdentityKey: {
-		object: { key: PUBLIC_KEY, type: SigningKeyPairType.Ed25519 },
-		version: 0,
-	},
-	sourceOfTrust: IdentityKeySourceOfTrust.Manual,
-}
 
-const TRUSTED_MAIL_ADDRESS = "trust-me@tuta.com"
-const PUBLIC_KEY_FINGERPRINT_SQL_RESULT: Record<string, TaggedSqlValue> = {
-	mailAddress: {
-		type: SqlType.String,
-		value: TRUSTED_MAIL_ADDRESS,
-	},
-	publicIdentityKey: {
-		type: SqlType.Bytes,
-		value: PUBLIC_KEY_BYTES,
-	},
-	identityKeyVersion: {
-		type: SqlType.Number,
-		value: 0,
-	},
-	identityKeyType: {
-		type: SqlType.Number,
-		value: SigningKeyPairType.Ed25519,
-	},
-	sourceOfTrust: {
-		type: SqlType.Number,
-		value: IdentityKeySourceOfTrust.Manual,
-	},
-}
+let trustDBEntry: TrustDBEntry
 
 o.spec("KeyVerificationFacadeTest", function () {
 	let keyVerification: KeyVerificationFacade
-	let sqlCipherFacade: SqlCipherFacade
 	let publicKeySignatureFacade: PublicKeySignatureFacade
-
-	let backupEnv: any
+	let publicIdentityKeyProvider: PublicIdentityKeyProvider
+	let identityKeyTrustDatabase: IdentityKeyTrustDatabase
+	let publicKeyIdentifier: PublicKeyIdentifier
+	let maybeSignedPublicKey: MaybeSignedPublicKey
 
 	o.beforeEach(function () {
-		sqlCipherFacade = object()
 		publicKeySignatureFacade = object()
+		publicIdentityKeyProvider = object()
+		identityKeyTrustDatabase = object()
 
-		keyVerification = new KeyVerificationFacade(sqlCipherFacade, publicKeySignatureFacade)
-	})
+		publicKeyIdentifier = {
+			identifier: object(),
+			identifierType: PublicKeyIdentifierType.MAIL_ADDRESS,
+		}
+		maybeSignedPublicKey = {
+			signature: createTestEntity(PublicKeySignatureTypeRef, { signature: object() }),
+			publicKey: object(),
+		}
+		trustDBEntry = {
+			publicIdentityKey: {
+				object: { key: PUBLIC_KEY, type: SigningKeyPairType.Ed25519 },
+				version: 0,
+			},
+			sourceOfTrust: IdentityKeySourceOfTrust.Manual,
+		}
 
-	o.spec("confirm trusted identity database works as intended", function () {
-		o("load only manually verified keys", async function () {
-			const sqlResult: Record<string, TaggedSqlValue>[] = []
-			when(sqlCipherFacade.all(anything(), anything())).thenResolve(sqlResult)
-
-			await keyVerification.getManuallyVerifiedIdentities()
-
-			verify(sqlCipherFacade.all("SELECT * FROM identity_store WHERE sourceOfTrust = 0", []))
-		})
-
-		o("identity database is empty", async function () {
-			const sqlResult: Record<string, TaggedSqlValue>[] = []
-			when(sqlCipherFacade.all(anything(), anything())).thenResolve(sqlResult)
-
-			const result = await keyVerification.getManuallyVerifiedIdentities()
-			const expectation = new Map<string, TrustedIdentity>()
-
-			o(result).deepEquals(expectation)
-		})
-
-		o("trusting an identity", async function () {
-			await keyVerification.trust(TRUSTED_MAIL_ADDRESS, PUBLIC_KEY_TRUSTED_IDENTITY.publicIdentityKey, IdentityKeySourceOfTrust.Manual)
-
-			const query =
-				"\n\t\t\tINSERT INTO identity_store (mailAddress, publicIdentityKey, identityKeyVersion, identityKeyType, sourceOfTrust)\n\t\t\tVALUES (?, ?, ?, ?, ?)"
-			const params: TaggedSqlValue[] = [
-				{
-					type: SqlType.String,
-					value: TRUSTED_MAIL_ADDRESS,
-				},
-				{
-					type: SqlType.Bytes,
-					value: PUBLIC_KEY_BYTES,
-				},
-				{
-					type: SqlType.Number,
-					value: 0,
-				},
-				{
-					type: SqlType.Number,
-					value: SigningKeyPairType.Ed25519,
-				},
-				{
-					type: SqlType.Number,
-					value: IdentityKeySourceOfTrust.Manual,
-				},
-			]
-			verify(sqlCipherFacade.run(query, params))
-		})
-
-		o("distrusting an identity", async function () {
-			await keyVerification.untrust("untrust-me@tuta.com")
-
-			const query = `DELETE FROM identity_store WHERE mailAddress = ?`
-			const params: TaggedSqlValue[] = [{ type: SqlType.String, value: "untrust-me@tuta.com" }]
-			verify(sqlCipherFacade.run(query, params))
-		})
-
-		o("checking for trust", async function () {
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(PUBLIC_KEY_FINGERPRINT_SQL_RESULT)
-			const trusted = await keyVerification.isTrusted(TRUSTED_MAIL_ADDRESS)
-
-			const query = `SELECT * FROM identity_store WHERE mailAddress = ?`
-			const params: TaggedSqlValue[] = [
-				{
-					type: SqlType.String,
-					value: TRUSTED_MAIL_ADDRESS,
-				},
-			]
-			verify(sqlCipherFacade.get(query, params))
-			o(trusted).equals(true)
-		})
+		keyVerification = new KeyVerificationFacade(publicKeySignatureFacade, publicIdentityKeyProvider, identityKeyTrustDatabase)
 	})
 
 	o.spec("verification state gets resolved correctly", function () {
-		o("manually trusted and verified keys result in VERIFIED_MANUAL", async function () {
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(PUBLIC_KEY_FINGERPRINT_SQL_RESULT)
-			when(publicKeySignatureFacade.verifyPublicKeySignature(anything(), anything(), anything())).thenResolve(true)
-			const encryptionPublicKey: Versioned<PublicKey> = object()
-			const encodedSignature: EncodedEd25519Signature = object()
-			const state = await keyVerification.verify(TRUSTED_MAIL_ADDRESS, encryptionPublicKey, encodedSignature)
-			o(state).equals(EncryptionKeyVerificationState.VERIFIED_MANUAL)
+		o("identifier type != MAILADRESS results in NOT_SUPPORTED", async function () {
+			publicKeyIdentifier.identifierType = PublicKeyIdentifierType.GROUP_ID
+			const result = await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey)
+			o(result.verificationState).equals(EncryptionKeyVerificationState.NOT_SUPPORTED)
+			o(result.publicEncryptionKey).equals(maybeSignedPublicKey.publicKey)
+			verify(publicIdentityKeyProvider.loadPublicIdentityKey(anything()), { times: 0 })
 		})
-
-		o("tofu trusted and verified keys result in VERIFIED_TOFU", async function () {
-			when(publicKeySignatureFacade.verifyPublicKeySignature(anything(), anything(), anything())).thenResolve(true)
-			const TOFU_SQL_RESULT: Record<string, TaggedSqlValue> = {}
-			Object.assign(TOFU_SQL_RESULT, PUBLIC_KEY_FINGERPRINT_SQL_RESULT, {
-				sourceOfTrust: {
-					type: SqlType.Number,
-					value: IdentityKeySourceOfTrust.TOFU,
-				},
-			})
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(TOFU_SQL_RESULT)
-
-			const encryptionPublicKey: Versioned<PublicKey> = object()
-			const encodedSignature: EncodedEd25519Signature = object()
-			const state = await keyVerification.verify(TRUSTED_MAIL_ADDRESS, encryptionPublicKey, encodedSignature)
-			o(state).equals(EncryptionKeyVerificationState.VERIFIED_TOFU)
+		o("no identity key, no signature result in NO_ENTRY", async function () {
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(null)
+			maybeSignedPublicKey.signature = null
+			const result = await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey)
+			o(result.verificationState).equals(EncryptionKeyVerificationState.NO_ENTRY)
+			o(result.publicEncryptionKey).equals(maybeSignedPublicKey.publicKey)
 		})
-
-		o("trusted but unverified keys result in MISMATCH", async function () {
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(PUBLIC_KEY_FINGERPRINT_SQL_RESULT)
-			when(publicKeySignatureFacade.verifyPublicKeySignature(anything(), anything(), anything())).thenResolve(false)
-
-			const encryptionPublicKey: Versioned<PublicKey> = object()
-			const encodedSignature: EncodedEd25519Signature = object()
-
-			await assertThrows(KeyVerificationMismatchError, async () => {
-				await keyVerification.verify(TRUSTED_MAIL_ADDRESS, encryptionPublicKey, encodedSignature)
-			})
+		o("no identity key, existing signature result in KeyVerificationMismatch", async function () {
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(null)
+			await assertThrows(KeyVerificationMismatchError, async () => await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey))
 		})
-
-		o("untrusted keys result in NO_ENTRY", async function () {
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(null)
-
-			const encryptionPublicKey: Versioned<PublicKey> = object()
-			const encodedSignature: EncodedEd25519Signature = object()
-
-			const state = await keyVerification.verify(TRUSTED_MAIL_ADDRESS, encryptionPublicKey, encodedSignature)
-			o(state).equals(EncryptionKeyVerificationState.NO_ENTRY)
-			verify(publicKeySignatureFacade.verifyPublicKeySignature(anything(), anything(), anything()), { times: 0 })
+		o("idenity key but no signature results in KeyVerificationMissmatch", async function () {
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(trustDBEntry)
+			maybeSignedPublicKey.signature = null
+			await assertThrows(KeyVerificationMismatchError, async () => await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey))
 		})
-
-		o("trusted keys without retreived signatures result in a mismatch error", async function () {
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(PUBLIC_KEY_FINGERPRINT_SQL_RESULT)
-
-			const encryptionPublicKey: Versioned<PublicKey> = object()
-			const encodedSignature: EncodedEd25519Signature | null = null
-
-			await assertThrows(KeyVerificationMismatchError, async () => {
-				await keyVerification.verify(TRUSTED_MAIL_ADDRESS, encryptionPublicKey, encodedSignature)
-			})
-			verify(publicKeySignatureFacade.verifyPublicKeySignature(anything(), anything(), anything()), { times: 0 })
+		o("invalid signature results in KeyVerificationMissmatch", async function () {
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(trustDBEntry)
+			when(
+				publicKeySignatureFacade.verifyPublicKeySignature(
+					maybeSignedPublicKey.publicKey,
+					trustDBEntry.publicIdentityKey.object,
+					maybeSignedPublicKey.signature!.signature,
+				),
+			).thenResolve(false)
+			await assertThrows(KeyVerificationMismatchError, async () => await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey))
+		})
+		o("valid signature returns correct verification state - manual", async function () {
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(trustDBEntry)
+			when(
+				publicKeySignatureFacade.verifyPublicKeySignature(
+					maybeSignedPublicKey.publicKey,
+					trustDBEntry.publicIdentityKey.object.key,
+					maybeSignedPublicKey.signature!.signature,
+				),
+			).thenResolve(true)
+			const result = await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey)
+			o(result.verificationState).equals(EncryptionKeyVerificationState.VERIFIED_MANUAL)
+			o(result.publicEncryptionKey).equals(maybeSignedPublicKey.publicKey)
+		})
+		o("valid signature returns correct verification state - tofu", async function () {
+			trustDBEntry.sourceOfTrust = IdentityKeySourceOfTrust.TOFU
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(trustDBEntry)
+			when(
+				publicKeySignatureFacade.verifyPublicKeySignature(
+					maybeSignedPublicKey.publicKey,
+					trustDBEntry.publicIdentityKey.object.key,
+					maybeSignedPublicKey.signature!.signature,
+				),
+			).thenResolve(true)
+			const result = await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey)
+			o(result.verificationState).equals(EncryptionKeyVerificationState.VERIFIED_TOFU)
+			o(result.publicEncryptionKey).equals(maybeSignedPublicKey.publicKey)
+		})
+		o("valid signature returns correct verification state - not supported", async function () {
+			trustDBEntry.sourceOfTrust = IdentityKeySourceOfTrust.Not_Supported
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(trustDBEntry)
+			when(
+				publicKeySignatureFacade.verifyPublicKeySignature(
+					maybeSignedPublicKey.publicKey,
+					trustDBEntry.publicIdentityKey.object.key,
+					maybeSignedPublicKey.signature!.signature,
+				),
+			).thenResolve(true)
+			const result = await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey)
+			o(result.verificationState).equals(EncryptionKeyVerificationState.NOT_SUPPORTED)
+			o(result.publicEncryptionKey).equals(maybeSignedPublicKey.publicKey)
 		})
 
 		o("non-Ed25519 identity keys result in an error", async function () {
-			const invalidTypeSqlResult: Record<string, TaggedSqlValue> = {}
-			Object.assign(invalidTypeSqlResult, PUBLIC_KEY_FINGERPRINT_SQL_RESULT, {
-				identityKeyType: {
-					type: SqlType.Number,
-					value: 10,
-				},
-			})
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(invalidTypeSqlResult)
-
-			const encryptionPublicKey: Versioned<PublicKey> = object()
-			const encodedSignature: EncodedEd25519Signature = object()
-
-			await assertThrows(ProgrammingError, async () => {
-				await keyVerification.verify(TRUSTED_MAIL_ADDRESS, encryptionPublicKey, encodedSignature)
-			})
-			verify(publicKeySignatureFacade.verifyPublicKeySignature(anything(), anything(), anything()), { times: 0 })
+			trustDBEntry.publicIdentityKey.object.type = 10 as SigningKeyPairType
+			when(publicIdentityKeyProvider.loadPublicIdentityKey(publicKeyIdentifier)).thenResolve(trustDBEntry)
+			when(
+				publicKeySignatureFacade.verifyPublicKeySignature(
+					maybeSignedPublicKey.publicKey,
+					trustDBEntry.publicIdentityKey.object,
+					maybeSignedPublicKey.signature!.signature,
+				),
+			).thenResolve(true)
+			await assertThrows(ProgrammingError, async () => await keyVerification.verify(publicKeyIdentifier, maybeSignedPublicKey))
 		})
 	})
 
-	o.spec("trusted identity acquisition", function () {
-		o("acquire existing identity from trust database", async function () {
-			when(sqlCipherFacade.get(anything(), anything())).thenResolve(PUBLIC_KEY_FINGERPRINT_SQL_RESULT)
-
-			const result = await keyVerification.getTrustedIdentity(TRUSTED_MAIL_ADDRESS)
-			o(result).deepEquals(PUBLIC_KEY_TRUSTED_IDENTITY)
-
-			const query = "SELECT * FROM identity_store WHERE mailAddress = ?"
-			const params: TaggedSqlValue[] = [{ type: SqlType.String, value: TRUSTED_MAIL_ADDRESS }]
-
-			verify(sqlCipherFacade.get(query, params))
+	o.spec("getManuallyVerifiedIdentities", function () {
+		let trustDbEntries: Map<string, TrustDBEntry>
+		const mailAddress = "trusted-entry@tuta.com"
+		o.beforeEach(function () {
+			trustDbEntries = new Map()
+			trustDbEntries.set(mailAddress, trustDBEntry)
+			when(identityKeyTrustDatabase.getManuallyVerifiedEntries()).thenResolve(trustDbEntries)
 		})
 
-		o("acquire non-existing identity from trust database", async function () {
-			const result = await keyVerification.getTrustedIdentity("missing@example.com")
-			o(result).equals(null)
-
-			const query = "SELECT * FROM identity_store WHERE mailAddress = ?"
-			const params: TaggedSqlValue[] = [{ type: SqlType.String, value: "missing@example.com" }]
-
-			verify(sqlCipherFacade.get(query, params))
+		o("success - fingerprint is added to trust db entries", async function () {
+			const result = await keyVerification.getManuallyVerifiedIdentities()
+			o(result.size).equals(1)
+			const trustIdentity = result.get(mailAddress)
+			o(trustIdentity?.sourceOfTrust).deepEquals(trustDBEntry.sourceOfTrust)
+			o(trustIdentity?.publicIdentityKey).deepEquals(trustDBEntry.publicIdentityKey)
+			o(trustIdentity?.fingerprint).deepEquals(PUBLIC_KEY_FINGERPRINT)
 		})
 	})
 
 	o.spec("fingerprint calculation is robust", function () {
 		o("generic fingerprint calculation works", async function () {
-			const fingerprint = await keyVerification.calculateFingerprint(PUBLIC_KEY_TRUSTED_IDENTITY.publicIdentityKey)
+			const fingerprint = keyVerification.calculateFingerprint(trustDBEntry.publicIdentityKey)
 			o(fingerprint).deepEquals(PUBLIC_KEY_FINGERPRINT)
 		})
 
@@ -275,23 +197,5 @@ o.spec("KeyVerificationFacadeTest", function () {
 			concatenation = keyVerification.concatenateFingerprint(publicKey)
 			verifyKeyMetadata(concatenation, 5, SigningKeyPairType.Ed25519)
 		})
-	})
-
-	o("feature should be supported when on desktop", async function () {
-		const isSupported = await withOverriddenEnv({ mode: Mode.Desktop }, () => keyVerification.isSupported())
-		o(isSupported).equals(true)
-	})
-
-	o("feature should NOT be supported when on browser", async function () {
-		const isSupported = await withOverriddenEnv({ mode: Mode.Browser }, () => keyVerification.isSupported())
-		o(isSupported).equals(false)
-	})
-
-	o("database should NOT be queried when key verification is not supported", async function () {
-		globalThis.env.mode = Mode.Browser
-		await assertThrows(ProgrammingError, async () => {
-			await keyVerification.verify(object(), object(), object())
-		})
-		verify(sqlCipherFacade.get(anything(), anything()), { times: 0 })
 	})
 })
